Added a Submitted By field to the PlantView. This is both a good and a bad thing.
This is good because when we browse plants in the system, we are able to view who submitted them.

This is a bad thing, because we are using identical views when browsing and adding... This causes an empty field in the add plant view, which cannot be populated... it looks bad. Solution:

Fork the view, so we have one for viewing and one for adding. Then on the add-plant view we can remove the field!

Also, there is a current bug which causes duplicate displaying of views on mobile.

Lastly, I added Sarah, Cole, and me to the team members on the page footer :)